+++ src/components/common/error/ErrorCode.ts
@@ -48,7 +48,11 @@ export const getErrorByCode = (error: AxiosError<{ status: number; name: string;
   const axiosErrorCode = error?.code ?? '';
 
   if (serverErrorCode && serverErrorCode.status in ERROR_CODE) {
-    return { ...ERROR_CODE[serverErrorCode.status], description: serverErrorCode.message[0] };
+    const englishPattern = /^[A-Za-z\s]+$/;
+
+    return englishPattern.test(serverErrorCode.message[0])
+      ? ERROR_CODE[serverErrorCode.status]
+      : { ...ERROR_CODE[serverErrorCode.status], description: serverErrorCode.message[0] };
   }
   if (httpErrorCode in ERROR_CODE) {
     return ERROR_CODE[httpErrorCode];
